+++ src/common/dsp/QuadFilterUnit.cpp
@@ -584,277 +584,6 @@ __m128 COMBquad_SSE2(QuadFilterUnitState *__restrict f, __m128 in)
     return _mm_add_ps(_mm_mul_ps(f->C[3], DBRead), _mm_mul_ps(f->C[2], in));
 }
 
-__m128 CLIP(QuadFilterWaveshaperState *__restrict s, __m128 in, __m128 drive)
-{
-    const __m128 x_min = _mm_set1_ps(-1.0f);
-    const __m128 x_max = _mm_set1_ps(1.0f);
-    return _mm_max_ps(_mm_min_ps(_mm_mul_ps(in, drive), x_max), x_min);
-}
-
-__m128 DIGI_SSE2(QuadFilterWaveshaperState *__restrict s, __m128 in, __m128 drive)
-{
-    // v1.2: return (double)((int)((double)(x*p0inv*16.f+1.0)))*p0*0.0625f;
-    const __m128 m16 = _mm_set1_ps(16.f);
-    const __m128 m16inv = _mm_set1_ps(0.0625f);
-    const __m128 mofs = _mm_set1_ps(0.5f);
-
-    __m128 invdrive = _mm_rcp_ps(drive);
-    __m128i a = _mm_cvtps_epi32(_mm_add_ps(mofs, _mm_mul_ps(invdrive, _mm_mul_ps(m16, in))));
-
-    return _mm_mul_ps(drive, _mm_mul_ps(m16inv, _mm_sub_ps(_mm_cvtepi32_ps(a), mofs)));
-}
-
-__m128 TANH(QuadFilterWaveshaperState *__restrict s, __m128 in, __m128 drive)
-{
-    // Closer to ideal than TANH0
-    // y = x * ( 27 + x * x ) / ( 27 + 9 * x * x );
-    // y = clip(y)
-    const __m128 m9 = _mm_set1_ps(9.f);
-    const __m128 m27 = _mm_set1_ps(27.f);
-
-    __m128 x = _mm_mul_ps(in, drive);
-    __m128 xx = _mm_mul_ps(x, x);
-    __m128 denom = _mm_add_ps(m27, _mm_mul_ps(m9, xx));
-    __m128 y = _mm_mul_ps(x, _mm_add_ps(m27, xx));
-    y = _mm_mul_ps(y, _mm_rcp_ps(denom));
-
-    const __m128 y_min = _mm_set1_ps(-1.0f);
-    const __m128 y_max = _mm_set1_ps(1.0f);
-    return _mm_max_ps(_mm_min_ps(y, y_max), y_min);
-}
-
-__m128 SINUS_SSE2(QuadFilterWaveshaperState *__restrict s, __m128 in, __m128 drive)
-{
-    const __m128 one = _mm_set1_ps(1.f);
-    const __m128 m256 = _mm_set1_ps(256.f);
-    const __m128 m512 = _mm_set1_ps(512.f);
-
-    __m128 x = _mm_mul_ps(in, drive);
-    x = _mm_add_ps(_mm_mul_ps(x, m256), m512);
-
-    __m128i e = _mm_cvtps_epi32(x);
-    __m128 a = _mm_sub_ps(x, _mm_cvtepi32_ps(e));
-    e = _mm_packs_epi32(e, e);
-    const __m128i UB = _mm_set1_epi16(0x3fe);
-    e = _mm_max_epi16(_mm_min_epi16(e, UB), _mm_setzero_si128());
-
-#if MAC
-    // this should be very fast on C2D/C1D (and there are no macs with K8's)
-    // GCC seems to optimize around the XMM -> int transfers so this is needed here
-    int e4 alignas(16)[4];
-    e4[0] = _mm_cvtsi128_si32(e);
-    e4[1] = _mm_cvtsi128_si32(_mm_shufflelo_epi16(e, _MM_SHUFFLE(1, 1, 1, 1)));
-    e4[2] = _mm_cvtsi128_si32(_mm_shufflelo_epi16(e, _MM_SHUFFLE(2, 2, 2, 2)));
-    e4[3] = _mm_cvtsi128_si32(_mm_shufflelo_epi16(e, _MM_SHUFFLE(3, 3, 3, 3)));
-#else
-    // on PC write to memory & back as XMM -> GPR is slow on K8
-    short e4 alignas(16)[8];
-    _mm_store_si128((__m128i *)&e4, e);
-#endif
-
-    __m128 ws1 = _mm_load_ss(&waveshapers[wst_sine][e4[0] & 0x3ff]);
-    __m128 ws2 = _mm_load_ss(&waveshapers[wst_sine][e4[1] & 0x3ff]);
-    __m128 ws3 = _mm_load_ss(&waveshapers[wst_sine][e4[2] & 0x3ff]);
-    __m128 ws4 = _mm_load_ss(&waveshapers[wst_sine][e4[3] & 0x3ff]);
-    __m128 ws = _mm_movelh_ps(_mm_unpacklo_ps(ws1, ws2), _mm_unpacklo_ps(ws3, ws4));
-    ws1 = _mm_load_ss(&waveshapers[wst_sine][(e4[0] + 1) & 0x3ff]);
-    ws2 = _mm_load_ss(&waveshapers[wst_sine][(e4[1] + 1) & 0x3ff]);
-    ws3 = _mm_load_ss(&waveshapers[wst_sine][(e4[2] + 1) & 0x3ff]);
-    ws4 = _mm_load_ss(&waveshapers[wst_sine][(e4[3] + 1) & 0x3ff]);
-    __m128 wsn = _mm_movelh_ps(_mm_unpacklo_ps(ws1, ws2), _mm_unpacklo_ps(ws3, ws4));
-
-    x = _mm_add_ps(_mm_mul_ps(_mm_sub_ps(one, a), ws), _mm_mul_ps(a, wsn));
-
-    return x;
-}
-
-__m128 ASYM_SSE2(QuadFilterWaveshaperState *__restrict s, __m128 in, __m128 drive)
-{
-    const __m128 one = _mm_set1_ps(1.f);
-    const __m128 m32 = _mm_set1_ps(32.f);
-    const __m128 m512 = _mm_set1_ps(512.f);
-    const __m128i UB = _mm_set1_epi16(0x3fe);
-
-    __m128 x = _mm_mul_ps(in, drive);
-    x = _mm_add_ps(_mm_mul_ps(x, m32), m512);
-
-    __m128i e = _mm_cvtps_epi32(x);
-    __m128 a = _mm_sub_ps(x, _mm_cvtepi32_ps(e));
-    e = _mm_packs_epi32(e, e);
-    e = _mm_max_epi16(_mm_min_epi16(e, UB), _mm_setzero_si128());
-
-#if MAC
-    // this should be very fast on C2D/C1D (and there are no macs with K8's)
-    int e4 alignas(16)[4];
-    e4[0] = _mm_cvtsi128_si32(e);
-    e4[1] = _mm_cvtsi128_si32(_mm_shufflelo_epi16(e, _MM_SHUFFLE(1, 1, 1, 1)));
-    e4[2] = _mm_cvtsi128_si32(_mm_shufflelo_epi16(e, _MM_SHUFFLE(2, 2, 2, 2)));
-    e4[3] = _mm_cvtsi128_si32(_mm_shufflelo_epi16(e, _MM_SHUFFLE(3, 3, 3, 3)));
-
-#else
-    // on PC write to memory & back as XMM -> GPR is slow on K8
-    short e4 alignas(16)[8];
-    _mm_store_si128((__m128i *)&e4, e);
-#endif
-
-    __m128 ws1 = _mm_load_ss(&waveshapers[wst_asym][e4[0] & 0x3ff]);
-    __m128 ws2 = _mm_load_ss(&waveshapers[wst_asym][e4[1] & 0x3ff]);
-    __m128 ws3 = _mm_load_ss(&waveshapers[wst_asym][e4[2] & 0x3ff]);
-    __m128 ws4 = _mm_load_ss(&waveshapers[wst_asym][e4[3] & 0x3ff]);
-    __m128 ws = _mm_movelh_ps(_mm_unpacklo_ps(ws1, ws2), _mm_unpacklo_ps(ws3, ws4));
-    ws1 = _mm_load_ss(&waveshapers[wst_asym][(e4[0] + 1) & 0x3ff]);
-    ws2 = _mm_load_ss(&waveshapers[wst_asym][(e4[1] + 1) & 0x3ff]);
-    ws3 = _mm_load_ss(&waveshapers[wst_asym][(e4[2] + 1) & 0x3ff]);
-    ws4 = _mm_load_ss(&waveshapers[wst_asym][(e4[3] + 1) & 0x3ff]);
-    __m128 wsn = _mm_movelh_ps(_mm_unpacklo_ps(ws1, ws2), _mm_unpacklo_ps(ws3, ws4));
-
-    x = _mm_add_ps(_mm_mul_ps(_mm_sub_ps(one, a), ws), _mm_mul_ps(a, wsn));
-
-    return x;
-}
-
-template <int xRes, int xCenter, int size>
-__m128 WS_LUT(QuadFilterWaveshaperState *__restrict s, const float *table, __m128 in, __m128 drive)
-{
-    const __m128 one = _mm_set1_ps(1.f);
-    const __m128 m32 = _mm_set1_ps(xRes);
-    const __m128 m512 = _mm_set1_ps(xCenter);
-    const __m128i UB = _mm_set1_epi16(size - 1);
-
-    __m128 x = _mm_mul_ps(in, drive);
-    x = _mm_add_ps(_mm_mul_ps(x, m32), m512);
-
-    __m128i e = _mm_cvtps_epi32(x);
-    __m128 a = _mm_sub_ps(x, _mm_cvtepi32_ps(e));
-    e = _mm_packs_epi32(e, e);
-    e = _mm_max_epi16(_mm_min_epi16(e, UB), _mm_setzero_si128());
-
-#if MAC
-    // this should be very fast on C2D/C1D (and there are no macs with K8's)
-    int e4 alignas(16)[4];
-    e4[0] = _mm_cvtsi128_si32(e);
-    e4[1] = _mm_cvtsi128_si32(_mm_shufflelo_epi16(e, _MM_SHUFFLE(1, 1, 1, 1)));
-    e4[2] = _mm_cvtsi128_si32(_mm_shufflelo_epi16(e, _MM_SHUFFLE(2, 2, 2, 2)));
-    e4[3] = _mm_cvtsi128_si32(_mm_shufflelo_epi16(e, _MM_SHUFFLE(3, 3, 3, 3)));
-
-#else
-    // on PC write to memory & back as XMM -> GPR is slow on K8
-    short e4 alignas(16)[8];
-    _mm_store_si128((__m128i *)&e4, e);
-#endif
-
-    __m128 ws1 = _mm_load_ss(&table[e4[0] & size]);
-    __m128 ws2 = _mm_load_ss(&table[e4[1] & size]);
-    __m128 ws3 = _mm_load_ss(&table[e4[2] & size]);
-    __m128 ws4 = _mm_load_ss(&table[e4[3] & size]);
-    __m128 ws = _mm_movelh_ps(_mm_unpacklo_ps(ws1, ws2), _mm_unpacklo_ps(ws3, ws4));
-    ws1 = _mm_load_ss(&table[(e4[0] + 1) & size]);
-    ws2 = _mm_load_ss(&table[(e4[1] + 1) & size]);
-    ws3 = _mm_load_ss(&table[(e4[2] + 1) & size]);
-    ws4 = _mm_load_ss(&table[(e4[3] + 1) & size]);
-    __m128 wsn = _mm_movelh_ps(_mm_unpacklo_ps(ws1, ws2), _mm_unpacklo_ps(ws3, ws4));
-
-    x = _mm_add_ps(_mm_mul_ps(_mm_sub_ps(one, a), ws), _mm_mul_ps(a, wsn));
-
-    return x;
-}
-
-__m128 DIODE_SIM(__m128 v)
-{
-    static constexpr float vlf = 0.1, vbf = -0.05;
-    static const auto vb = _mm_set1_ps(vbf), vl = _mm_set1_ps(vlf), z = _mm_setzero_ps();
-
-    auto vLessVb = _mm_cmplt_ps(v, vb);
-    auto vBetweelVlVb = _mm_andnot_ps(vLessVb, _mm_cmplt_ps(v, vl)); // !A && B
-    auto vAboveVl = _mm_cmpge_ps(v, vl);
-
-    static const auto betweenDen = _mm_set1_ps(0.5f / (vlf - vbf)), vlvb = _mm_set1_ps(vlf - vbf),
-                      aboveConstant = _mm_set1_ps(0.5f * (vlf - vbf) * (vlf - vbf) / (vlf - vbf));
-    auto vvb = _mm_sub_ps(v, vb);
-    auto vlvbVal = _mm_mul_ps(_mm_mul_ps(vvb, vvb), betweenDen);
-    auto vaboveVal = _mm_add_ps(_mm_sub_ps(v, vl), aboveConstant);
-    auto r = _mm_add_ps(
-        z, _mm_add_ps(_mm_and_ps(vBetweelVlVb, vlvbVal), _mm_and_ps(vAboveVl, vaboveVal)));
-    return r;
-}
-
-__m128 FULL_WAVE(QuadFilterWaveshaperState *__restrict s, __m128 x, __m128 drive)
-{
-    static const auto m1 = _mm_set1_ps(-1.0f);
-
-    // auto aps = abs_ps(x);
-    x = CLIP(s, x, drive);
-    auto aps = DIODE_SIM(x);
-    auto apn = DIODE_SIM(_mm_mul_ps(x, m1));
-    return _mm_add_ps(aps, apn);
-}
-
-__m128 POS_WAVE(QuadFilterWaveshaperState *__restrict s, __m128 x, __m128 drive)
-{
-    // auto aps = abs_ps(x);
-    x = CLIP(s, x, drive);
-    return DIODE_SIM(x);
-}
-
-__m128 NEG_WAVE(QuadFilterWaveshaperState *__restrict s, __m128 x, __m128 drive)
-{
-    static const auto m1 = _mm_set1_ps(-1.0f);
-
-    // auto aps = abs_ps(x);
-    x = CLIP(s, x, drive);
-    return DIODE_SIM(_mm_mul_ps(x, m1));
-}
-
-template <__m128 (*K)(__m128)>
-__m128 CHEBY_CORE(QuadFilterWaveshaperState *__restrict s, __m128 x, __m128 drive)
-{
-    static const auto m1 = _mm_set1_ps(-1.0f);
-    static const auto p1 = _mm_set1_ps(1.0f);
-
-    auto bound = K(_mm_max_ps(_mm_min_ps(x, p1), m1));
-    return TANH(s, bound, drive);
-}
-
-__m128 cheb2_kernel(__m128 x) // 2 x^2 - 1
-{
-    static const auto m1 = _mm_set1_ps(1);
-    static const auto m2 = _mm_set1_ps(2);
-    return _mm_sub_ps(_mm_mul_ps(m2, _mm_mul_ps(x, x)), m1);
-}
-
-__m128 cheb3_kernel(__m128 x) // 4 x^3 - 3 x
-{
-    static const auto m4 = _mm_set1_ps(4);
-    static const auto m3 = _mm_set1_ps(3);
-    auto x2 = _mm_mul_ps(x, x);
-    auto v4x2m3 = _mm_sub_ps(_mm_mul_ps(m4, x2), m3);
-    return _mm_mul_ps(v4x2m3, x);
-}
-
-__m128 cheb4_kernel(__m128 x) // 8 x^4 - 8 x^2 + 1
-{
-    static const auto m1 = _mm_set1_ps(1);
-    static const auto m8 = _mm_set1_ps(8);
-    auto x2 = _mm_mul_ps(x, x);
-    auto x4mx2 = _mm_mul_ps(x2, _mm_sub_ps(x2, m1)); // x^2 * ( x^2 - 1)
-    return _mm_add_ps(_mm_mul_ps(m8, x4mx2), m1);
-}
-
-__m128 cheb5_kernel(__m128 x) // 16 x^5 - 20 x^3 + 5 x
-{
-    static const auto m16 = _mm_set1_ps(16);
-    static const auto mn20 = _mm_set1_ps(-20);
-    static const auto m5 = _mm_set1_ps(5);
-
-    auto x2 = _mm_mul_ps(x, x);
-    auto x3 = _mm_mul_ps(x2, x);
-    auto x5 = _mm_mul_ps(x3, x2);
-    auto t1 = _mm_mul_ps(m16, x5);
-    auto t2 = _mm_mul_ps(mn20, x3);
-    auto t3 = _mm_mul_ps(m5, x);
-    return _mm_add_ps(t1, _mm_add_ps(t2, t3));
-}
-
 FilterUnitQFPtr GetQFPtrFilterUnit(int type, int subtype)
 {
     // Force compiler to error out if I miss one
@@ -999,41 +728,6 @@ FilterUnitQFPtr GetQFPtrFilterUnit(int type, int subtype)
     case fut_none:
     case n_fu_types:
         return 0;
-    }
-    return 0;
-}
-
-WaveshaperQFPtr GetQFPtrWaveshaper(int type)
-{
-    switch (type)
-    {
-    case wst_soft:
-        return TANH;
-    case wst_hard:
-        return CLIP;
-    case wst_asym:
-        // return ASYM_SSE2;
-        return [](QuadFilterWaveshaperState *__restrict s, auto a, auto b) {
-            return WS_LUT<32, 512, 0x3ff>(s, waveshapers[wst_asym], a, b);
-        };
-    case wst_sine:
-        return SINUS_SSE2;
-    case wst_digital:
-        return DIGI_SSE2;
-    case wst_cheby2:
-        return CHEBY_CORE<cheb2_kernel>;
-    case wst_cheby3:
-        return CHEBY_CORE<cheb3_kernel>;
-    case wst_cheby4:
-        return CHEBY_CORE<cheb4_kernel>;
-    case wst_cheby5:
-        return CHEBY_CORE<cheb5_kernel>;
-    case wst_fwrectify:
-        return FULL_WAVE;
-    case wst_poswav:
-        return POS_WAVE;
-    case wst_negwav:
-        return NEG_WAVE;
     }
     return 0;
 }
