+++ src/common/dsp/QuadFilterUnit.h
@@ -5,6 +5,9 @@
 const int n_filter_registers = 16;
 const int n_waveshaper_registers = 4;
 
+/*
+ * These are defined in QuadFilterUnit.cpp
+ */
 struct alignas(16) QuadFilterUnitState
 {
     __m128 C[n_cm_coeffs], dC[n_cm_coeffs]; // coefficients
@@ -13,17 +16,8 @@ struct alignas(16) QuadFilterUnitState
     int active[4]; // 0xffffffff if voice is active, 0 if not (usable as mask)
     int WP[4];     // comb write position
 };
-
-struct alignas(16) QuadFilterWaveshaperState
-{
-    __m128 R[n_waveshaper_registers];
-};
-
 typedef __m128 (*FilterUnitQFPtr)(QuadFilterUnitState *__restrict, __m128 in);
-typedef __m128 (*WaveshaperQFPtr)(QuadFilterWaveshaperState *__restrict, __m128 in, __m128 drive);
-
 FilterUnitQFPtr GetQFPtrFilterUnit(int type, int subtype);
-WaveshaperQFPtr GetQFPtrWaveshaper(int type);
 
 /*
  * Subtypes are integers below 16 - maybe one day go as high as 32. So we have space in the
@@ -40,3 +34,18 @@ enum QFUSubtypeMasks : int32_t
     UNMASK_SUBTYPE = (1 << 8) - 1,
     EXTENDED_COMB = 1 << 9
 };
+
+/*
+ * These are defined in QuadFilterWaveshapers.cpp
+ */
+struct alignas(16) QuadFilterWaveshaperState
+{
+    __m128 R[n_waveshaper_registers];
+};
+typedef __m128 (*WaveshaperQFPtr)(QuadFilterWaveshaperState *__restrict, __m128 in, __m128 drive);
+WaveshaperQFPtr GetQFPtrWaveshaper(int type);
+/*
+ * Given the very first sample inbound to a new voice session, return the
+ * first set of registers for that voice.
+ */
+void initializeWaveshaperRegister(int type, float R[n_waveshaper_registers]);
